Bug 1250356: rollup of changes for webrtc after applying webrtc.org v49 update r=pkerr,ng,pehrsons,etc See ssh://hg.mozilla.org/users/paulrkerr_gmail.com/webrtc49_merge/ for the patch development history.
UltraBlame original commit: 126348e718d03dec640b30b5def70fce8aa71527

+++ media/webrtc/trunk/webrtc/modules/video_coding/content_metrics_processing.h
@@ -41,7 +41,7 @@ class VCMContentMetricsProcessing {
   int Reset();
 
   
-  void UpdateFrameRate(uint32_t frameRate);
+  void UpdateFrameRate(float frameRate);
